# public/openapi.yaml
openapi: 3.0.3
info:
  title: Abel Lumber Builder Platform API
  description: |
    The Abel OS Builder Platform Core API for builders and staff.

    **Authentication Methods:**
    - **Builder Auth**: `abel_session` httpOnly cookie (set via POST /api/auth/login)
    - **Staff Auth**: Headers `x-staff-id`, `x-staff-role`, `x-staff-email` (set by middleware)

    **Rate Limiting**:
    - Standard auth endpoints: 10 requests/minute per IP
    - Public endpoints: 5 requests/minute per IP
    - Response headers: `X-RateLimit-Limit`, `X-RateLimit-Remaining`, `X-RateLimit-Reset`

    **Standard Error Response**:
    ```json
    {
      "error": "error_code",
      "message": "optional detailed message"
    }
    ```
  version: 1.0.0
  contact:
    name: Abel Lumber Platform Team
    url: https://app.abellumber.com
servers:
  - url: https://app.abellumber.com/api
    description: Production
  - url: http://localhost:3000/api
    description: Local development
...
paths:
...
  /builders/register:
    post:
      tags:
        - Builder Registration
      summary: Public builder registration
      description: Submit new builder application. Creates builder account with PENDING status for admin approval.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - companyName
                - contactName
                - contactEmail
              properties:
                companyName:
                  type: string
                  minLength: 1
                contactName:
                  type: string
                  minLength: 1
                contactEmail:
                  type: string
                  format: email
                contactPhone:
                  type: string
                address:
                  type: string
                city:
                  type: string
                state:
                  type: string
                zip:
                  type: string
                businessLicense:
                  type: string
                taxId:
                  type: string
                estimatedAnnualVolume:
                  type: string
                referralSource:
                  type: string
                notes:
                  type: string
      responses:
        '201':
          description: Application submitted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  refNumber:
                    type: string
                    description: Application reference number
                  applicationId:
                    type: string
        '400':
          description: Missing required fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
        '500':
          description: Server error
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Error code or message
        message:
          type: string
          description: Optional detailed error message
...
  securitySchemes:
    builderSession:
      type: apiKey
      name: abel_session
      in: cookie
      description: Builder session cookie (set via POST /api/auth/login)

    staffHeaders:
      type: apiKey
      name: x-staff-id
      in: header
      description: Staff authentication via headers (x-staff-id, x-staff-role, x-staff-email)